move breakWord into decompose
